[PM-26974] Add reprompt check when deleting ciphers in browser extension (#16907)
* Add reprompt check when deleting ciphers in browser extension

* RE-use existing remprompt check function.

+++ apps/browser/src/vault/popup/components/vault-v2/item-more-options/item-more-options.component.ts
@@ -267,6 +267,11 @@ export class ItemMoreOptionsComponent {
   }
 
   protected async delete() {
+    const repromptPassed = await this.passwordRepromptService.passwordRepromptCheck(this.cipher);
+    if (!repromptPassed) {
+      return;
+    }
+
     const confirmed = await this.dialogService.openSimpleDialog({
       title: { key: "deleteItem" },
       content: { key: "deleteItemConfirmation" },
